+++ data_processors/pipeline/orchestration/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""orchestration package
+
+That compliment Orchestrator act upon a particular event and/or condition to PERFORM some step(s).
+Typically contain X_step modules file that implement perform() interface for a particular step in the pipeline.
+"""
