+++ data_processors/pipeline/tests/test_sqs_iap_event.py
@@ -9,7 +9,7 @@ from data_portal.models import GDSFile, SequenceRun, Workflow, BatchRun, LabMeta
     LabMetadataWorkflow
 from data_portal.tests.factories import GDSFileFactory, WorkflowFactory, TestConstant
 from data_processors.pipeline.constant import WorkflowStatus, WorkflowType, WorkflowRunEventType
-from data_processors.pipeline.lambdas import sqs_iap_event, bcl_convert, update_google_lims
+from data_processors.pipeline.lambdas import sqs_iap_event, bcl_convert
 from data_processors.pipeline.tests import _rand, _uuid
 from data_processors.pipeline.tests.case import logger, PipelineUnitTestCase
 from utils import libslack, libjson
@@ -217,8 +217,6 @@ class SQSIAPEventUnitTests(PipelineUnitTestCase):
                 "PTC_EXPn200908LL_L2000001_topup"
             ]
         )
-        # ignore the google lims update (that's covered elsewhere)
-        when(update_google_lims).update_google_lims(any).thenReturn(True)
 
     def test_uploaded_gds_file_event(self):
         """
