+++ data_processors/pipeline/tools/tests/test_liborca.py
@@ -1,5 +1,6 @@
 import json
 
+from data_portal.tests.factories import TestConstant
 from data_processors.pipeline.tools import liborca
 from data_processors.pipeline.tests.case import PipelineUnitTestCase, logger
 
@@ -49,3 +50,17 @@ class LibOrcaUnitTests(PipelineUnitTestCase):
         except Exception as e:
             logger.exception(f"THIS ERROR EXCEPTION IS INTENTIONAL FOR TEST. NOT ACTUAL ERROR. \n{e}")
         self.assertRaises(json.JSONDecodeError)
+
+    def test_get_run_number_from_run_name(self):
+        """
+        python manage.py test data_processors.pipeline.tools.tests.test_liborca.LibOrcaUnitTests.test_get_run_number_from_run_name
+        """
+        run_no = liborca.get_run_number_from_run_name(TestConstant.sqr_name.value)
+        self.assertEqual(run_no, 1)
+
+    def test_get_timestamp_from_run_name(self):
+        """
+        python manage.py test data_processors.pipeline.tools.tests.test_liborca.LibOrcaUnitTests.test_get_timestamp_from_run_name
+        """
+        run_date = liborca.get_timestamp_from_run_name(TestConstant.sqr_name.value)
+        self.assertEqual(run_date, "2020-05-08")
